--- projects/config/onx/constant.js
@@ -1,23 +1,0 @@
-const tokenAddresses = {
-  onx: "0xe0ad1806fd3e7edf6ff52fdb822432e847411033",
-  sOnx: "0xa99f0ad2a539b2867fcfea47f7e71f240940b47c",
-  onxFarm: "0x168f8469ac17dd39cd9a2c2ead647f814a488ce9",
-  onxTripleFarm: "0x30D1A19EA928cCf46634cBC2944D2D89Be636f22",
-  onxWethSushiPair: "0x62f22A47e5D2F8b71cC44fD85863753618312f67",
-  pool: "0x47F3E6C1Ef0cBe69502167095b592e61de108BaA",
-  aethToken: "0xE95A203B1a91a908F9B9CE46459d101078c2c3cb",
-  onsToken: "0xfC97C0c12438B6E4CF246cD831b02FeF4950DCAD",
-  oneVault: "0x3BdF1977d87EDAD8e0617efCEa958F6d43A4C30E",
-  onePools: [
-      '0x54c532E367031F56c401C6024aC4ABEfF2b03534',
-      '0xbc2346C3c7F3998A12A1c8E7Be44734EcC832763',
-      '0xe64796FCc97c33A2193Ba60f013F3fA5D4712d56',
-      '0xe659fA84e0C687760245046BA63329d44320997c',
-      '0x19Eb6536777713aCdAcA8dd9A3AD9843D74E9E3b',
-      '0x1B72255a11DF705Bb598E670830E03A19F2D242f',
-      '0x7058EE5467edef417746aab0B8cabbbE36eF1798',
-      '0x9792eE4c36a622a8CF9566b037c57519A9fe8a56',
-  ]
-};
-
-module.exports = tokenAddresses;
