--- projects/config/onx/vault.js
@@ -1,13 +1,0 @@
-const EXCHANGE_TYPE = {
-  SUSHISWAP: 'SUSHISWAP',
-  AVALANCHE: 'AVALANCHE',
-  UNISWAP: 'UNISWAP',
-  QUICKSWAP: 'QUICKSWAP',
-  QUICKSWAP_DUAL_MATIC: 'QUICKSWAP_DUAL_MATIC',
-  SPOOKYSWAP: 'SPOOKYSWAP',
-  SPOOKYSWAP_SINGLE: 'SPOOKYSWAP_SINGLE',
-}
-
-module.exports = {
-  EXCHANGE_TYPE,
-}
