--- projects/config/piedao/abi/IStakingAll.json
@@ -1,4 +1,0 @@
-{
-  "poolCount": "uint256:poolCount",
-  "getPoolToken": "function getPoolToken(uint256 _poolId) view returns (address)"
-}
